Enhance [Internal] [Authentication] Cryptography Techniques
- [+] feat(crypto): add encryption and decryption functions using a cascade of ciphers (AES and ChaCha20-Poly1305)
- [+] refactor(crypto): split encryption and decryption logic into separate files for better organization
- [+] feat(crypto): introduce higher-order functions for encrypting and decrypting data using provided ciphers
- [+] feat(crypto): add support for encrypting and decrypting large data using AES encryption
- [+] refactor(crypto): move helper functions to a separate file for better code organization
- [+] feat(crypto): add function to verify the integrity of ciphertext without decrypting it
- [+] docs(crypto): add copyright notice and license information to all files
- [+] docs(crypto): add comments explaining the purpose and usage of each function
- [+] refactor(crypto): rename encrypt_test.go to crypto_test.go for consistency
- [+] refactor(crypto): remove unnecessary blank lines and improve code formatting
